+++ CHANGELOG.md
@@ -11,6 +11,8 @@
     - Use this function can pop val from the list
 - Add new function for `hm_map` -- `hm_map_pop()`
     - Use this function can pop entry from the list
+- Add new function for `hm_set` -- `hm_set_pop()`
+    - Use this function can pop entry from the set
 
 ### Changed
 - **refactor:** Change the return content of function `hm_map_get()`, using the new function `hm_map_get_entry()` to replace it